tests/requirements.txt
selenium>=4.15.0
pytest>=7.4.0
requests>=2.31.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
